+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Automatic title generation
 - [frontend] - Collapsible side bar.
 - [frontend] - Add confirmation box on thread deletion.
+- Feedback for HIL tool call refusal.
 
 ### Fixed
 - [frontend] - Parallel execution of tools